Create a class to convert plugin data;
PWPA-1937
