+++ src/alfasim_score/converter/alfacase/convert_alfacase.py
@@ -24,6 +24,7 @@ from alfasim_sdk import XAndYDescription
 from barril.units import Scalar
 
 from alfasim_score.common import convert_quota_to_tvd
+from alfasim_score.common import filter_duplicated_materials_by_name
 from alfasim_score.constants import CASING_DEFAULT_ROUGHNESS
 from alfasim_score.constants import CEMENT_NAME
 from alfasim_score.constants import FLUID_DEFAULT_NAME
@@ -38,14 +39,6 @@ from alfasim_score.constants import WELLBORE_TOP_NODE_NAME
 from alfasim_score.converter.alfacase.score_input_reader import ScoreInputReader
 from alfasim_score.units import LENGTH_UNIT
 from alfasim_score.units import TEMPERATURE_UNIT
-
-
-def filter_duplicated_materials(
-    material_list: List[MaterialDescription],
-) -> List[MaterialDescription]:
-    """Remove the duplicated materials parsed by the reader."""
-    filtered = {material.name: material for material in material_list}
-    return list(filtered.values())
 
 
 def get_section_top_of_filler(
@@ -88,7 +81,7 @@ class ScoreAlfacaseConverter:
             + self.score_input.read_lithology_materials()
             + self.score_input.read_packer_fluid()
         )
-        for material in material_list:
+        for material in filter_duplicated_materials_by_name(material_list):
             material_descriptions.append(
                 MaterialDescription(
                     name=material["name"],
@@ -99,7 +92,7 @@ class ScoreAlfacaseConverter:
                     expansion=material["thermal_expansion"],
                 )
             )
-        return filter_duplicated_materials(material_descriptions)
+        return material_descriptions
 
     def _convert_formation(self) -> FormationDescription:
         """Create the description for the formations."""
